split out gui testing in own file
